pyproject.toml
[project]
name = "fit-pal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.8",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "aiosqlite>=0.19.0",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.12",
    "langsmith-fetch>=0.3.1",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "anyio>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "pandas-stubs>=3.0.0.260204",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
